refactor: Extracts routing utilites to a separate package
