+++ plugins/governance/routing.go
@@ -3,7 +3,6 @@ package governance
 import (
 	"fmt"
 	"math/rand/v2"
-	"regexp"
 	"strings"
 
 	"github.com/google/cel-go/cel"
@@ -13,18 +12,6 @@ import (
 
 // DefaultRoutingChainMaxDepth is the default maximum depth for routing rule chain evaluation.
 const DefaultRoutingChainMaxDepth = 10
-
-// headerKeyPattern matches header map access patterns like headers["X-Api-Key"] or headers['X-Api-Key']
-var headerKeyPattern = regexp.MustCompile(`headers\[["']([^"']+)["']\]`)
-
-// headerInPattern matches "in headers" membership test patterns like "X-Api-Key" in headers or 'X-Api-Key' in headers
-var headerInPattern = regexp.MustCompile(`["']([^"']+)["']\s+in\s+headers`)
-
-// paramKeyPattern matches param map access patterns like params["Region"] or params['Region']
-var paramKeyPattern = regexp.MustCompile(`params\[["']([^"']+)["']\]`)
-
-// paramInPattern matches "in params" membership test patterns like "Region" in params or 'Region' in params
-var paramInPattern = regexp.MustCompile(`["']([^"']+)["']\s+in\s+params`)
 
 // ScopeLevel represents a level in the scope precedence hierarchy
 type ScopeLevel struct {
@@ -480,52 +467,6 @@ func scopeChainToStrings(chain []ScopeLevel) []string {
 		}
 	}
 	return scopes
-}
-
-// validateCELExpression performs basic validation on CEL expression format
-func validateCELExpression(expr string) error {
-	if expr == "" || expr == "true" || expr == "false" {
-		return nil // Empty, true, or false are valid
-	}
-
-	// List of allowed operators and keywords
-	validPatterns := []string{
-		"==", "!=", "&&", "||", ">", "<", ">=", "<=",
-		"in ", "matches ", ".startsWith(", ".contains(", ".endsWith(",
-		"[", "]", "(", ")", "!",
-	}
-
-	// Check if expression contains at least one valid operator
-	hasPattern := false
-	for _, pattern := range validPatterns {
-		if strings.Contains(expr, pattern) {
-			hasPattern = true
-			break
-		}
-	}
-
-	if !hasPattern {
-		return fmt.Errorf("expression must contain at least one operator: %s", expr)
-	}
-
-	return nil
-}
-
-// normalizeMapKeysInCEL lowercases header and param keys in CEL expressions
-// so that headers["X-Api-Key"] becomes headers["x-api-key"], "X-Api-Key" in headers becomes "x-api-key" in headers,
-// params["Region"] becomes params["region"], and "Region" in params becomes "region" in params.
-// This ensures CEL expressions match against the normalized (lowercase) map keys at runtime.
-func normalizeMapKeysInCEL(expr string) string {
-	toLower := func(match string) string {
-		return strings.ToLower(match)
-	}
-	// Normalize bracket access
-	expr = headerKeyPattern.ReplaceAllStringFunc(expr, toLower)
-	expr = paramKeyPattern.ReplaceAllStringFunc(expr, toLower)
-	// Normalize "in" membership test
-	expr = headerInPattern.ReplaceAllStringFunc(expr, toLower)
-	expr = paramInPattern.ReplaceAllStringFunc(expr, toLower)
-	return expr
 }
 
 // createCELEnvironment creates a new CEL environment for routing rules
